+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/execution_components.cpp
@@ -1,9 +1,9 @@
 #include "barretenberg/vm2/simulation/gadgets/execution_components.hpp"
+
 #include "barretenberg/vm2/simulation/events/addressing_event.hpp"
 #include "barretenberg/vm2/simulation/events/event_emitter.hpp"
 #include "barretenberg/vm2/simulation/gadgets/addressing.hpp"
 #include "barretenberg/vm2/simulation/gadgets/context.hpp"
-#include "barretenberg/vm2/simulation/standalone/pure_addressing.hpp"
 
 namespace bb::avm2::simulation {
 
@@ -18,18 +18,6 @@ std::unique_ptr<AddressingInterface> ExecutionComponentsProvider::make_addressin
 std::unique_ptr<GasTrackerInterface> ExecutionComponentsProvider::make_gas_tracker(GasEvent& gas_event,
                                                                                    const Instruction& instruction,
                                                                                    ContextInterface& context)
-{
-    return std::make_unique<GasTracker>(gas_event, instruction, instruction_info_db, context, greater_than);
-}
-
-std::unique_ptr<AddressingInterface> PureExecutionComponentsProvider::make_addressing(AddressingEvent&)
-{
-    return std::make_unique<PureAddressing>(instruction_info_db);
-}
-
-std::unique_ptr<GasTrackerInterface> PureExecutionComponentsProvider::make_gas_tracker(GasEvent& gas_event,
-                                                                                       const Instruction& instruction,
-                                                                                       ContextInterface& context)
 {
     return std::make_unique<GasTracker>(gas_event, instruction, instruction_info_db, context, greater_than);
 }
