+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/execution_components.hpp
@@ -43,23 +43,4 @@ class ExecutionComponentsProvider : public ExecutionComponentsProviderInterface 
     std::vector<std::unique_ptr<EventEmitterInterface<AddressingEvent>>> addressing_event_emitters;
 };
 
-class PureExecutionComponentsProvider : public ExecutionComponentsProviderInterface {
-  public:
-    PureExecutionComponentsProvider(GreaterThanInterface& greater_than,
-                                    const InstructionInfoDBInterface& instruction_info_db)
-        : greater_than(greater_than)
-        , instruction_info_db(instruction_info_db)
-    {}
-    std::unique_ptr<AddressingInterface> make_addressing(AddressingEvent& event) override;
-
-    // FIXME: not pure yet.
-    std::unique_ptr<GasTrackerInterface> make_gas_tracker(GasEvent& gas_event,
-                                                          const Instruction& instruction,
-                                                          ContextInterface& context) override;
-
-  private:
-    GreaterThanInterface& greater_than;
-    const InstructionInfoDBInterface& instruction_info_db;
-};
-
 } // namespace bb::avm2::simulation
